Refactor some functionalities to bootstrap the application (#92)
* Add CompanyController and CompanyService, change entities


* Add CompanyController and CompanyService, change entities


* Disable Spring Boot Security for a while


* Redesign some classes


* Refactor EmployerRelationshipService implementation


* Add post feature components (controller, service, dto)

+++ src/main/java/com/mangalaxy/mango/repository/TopicRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TopicRepository extends JpaRepository<Topic, Short> {
+public interface TopicRepository extends JpaRepository<Topic, Integer> {
 
 }
